+++ src/extension.ts
@@ -1,26 +1,15 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
+import { generateComponent } from '@commonCapabilities';
 import * as vscode from 'vscode';
 
-// This method is called when your extension is activated
-// Your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
+	const generateComponentDisposable = vscode.commands.registerCommand(
+		'gdlc-angular-toolbox.common-capabilities.generate-component',
+		async (arg: vscode.Uri) => {
+			await generateComponent(arg.fsPath);
+		},
+	);
 
-	// Use the console to output diagnostic information (console.log) and errors (console.error)
-	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "gdlc-angular-extension" is now active!');
-
-	// The command has been defined in the package.json file
-	// Now provide the implementation of the command with registerCommand
-	// The commandId parameter must match the command field in package.json
-	const disposable = vscode.commands.registerCommand('gdlc-angular-extension.helloWorld', () => {
-		// The code you place here will be executed every time your command is executed
-		// Display a message box to the user
-		vscode.window.showInformationMessage('Hello World from Angular Toolboox!');
-	});
-
-	context.subscriptions.push(disposable);
+	context.subscriptions.push(generateComponentDisposable);
 }
 
-// This method is called when your extension is deactivated
 export function deactivate() {}