+++ CHANGELOG.md
@@ -6,4 +6,33 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+### Added
+
+- Generate Component ✅
+  - With custom template ✅
+  - With custom spec template ✅
+  - Ask if selector prefix is needed or not ✅
+  - Ask if spec should be generated or not ✅
+- Generate Service 🔄
+  - With custom HTTP and none HTTP template 🔄
+  - With custom HTTP and none HTTP spec template 🔄
+  - Ask if spec should be generated or not 🔄
+  - Ask if service should global or not 🔄
+  - Ask if the service is an HTTP service or not 🔄
+- Generate Pipe
+  - With custom template 🔄
+  - With custom spec template 🔄
+  - Ask if spec should be generated or not 🔄
+  - Ask if pipe should be pure 🔄
+- Generate Directive
+  - With custom template 🔄
+  - With custom spec template 🔄
+  - Ask if spec should be generated or not 🔄
+  - Ask if the directive should have an alias 🔄
+- Configurations 🔄
+  - Set component prefix 🔄
+  - Set Angular version 🔄
+
+## [Released]
+
+## [0.1.0] -